Now outputs 1 viable solution for angles on the interval [-180,180]
Does not however work for the constraints of sin(theta4+theta3+delta)=0, nor cos(theta1+theta2)=0 although the ladder can be easily solved by adding an if statement and using the other calculated equation in terms of sin when this condition occurs. 

Also, note that in practice both constraints can easily be avoided by approximating the input precision whenever these theta combinations are encountered, on the extreme off chance that they are computed.